relocate hook
